challenge class and less global state
