+++ agent_or1.py
@@ -1,123 +1,62 @@
 from agentshells.shell1 import AgentShell
-
 import os
-import requests
 import json
 import re
+import sys
+import argparse
+from config import *
+from challenge import get_challenge_by_id, Challenge
+from utils import extract_xml_structures, get_completion
 
-challenges_dir = '/home/ec2-user/projects/unixagentbench/challenges/'
-challenge_id = '001'
-challenge_dir = os.path.join(challenges_dir, challenge_id)
-challenge_workdir = os.path.join(challenge_dir, 'workdir')
-assert os.path.exists(challenge_dir), f'Challenge dir {challenge_dir} does not exist'
-assert os.path.exists(challenge_workdir), f'Challenge workdir {challenge_workdir} does not exist'
-task_prompt_txt_file_path = os.path.join(challenge_dir, 'task_prompt.txt')
-assert os.path.exists(task_prompt_txt_file_path), f'Task prompt file {task_prompt_txt_file_path} does not exist'
-task_prompt = open(task_prompt_txt_file_path).read().strip()
-
-'''
-openai/gpt-4-turbo-preview
-openai/gpt-3.5-turbo
-01-ai/yi-34b-chat
-google/gemini-pro
-mistralai/mixtral-8x7b
-'''
-model = 'openai/gpt-3.5-turbo' #'mistralai/mixtral-8x7b'
-
-# get OPENROUTER_API_KEY from env
-OPENROUTER_API_KEY = os.getenv('OPENROUTER_API_KEY')
-print(OPENROUTER_API_KEY)
-
-def extract_xml_structures(input_string):
-    pattern = r'<(\w+)>(.+?)</\1>'
-    matches = re.findall(pattern, input_string)
-    
-    result = []
-    for match in matches:
-        tag, content = match
-        element = {
-            'tag': tag,
-            'content': content.strip()
-        }
-        result.append(element)
-    
-    return result
-
-messages = [
-    {
-        "role": "user",
-        "content": f"You are a friendly chatbot that has access to a stateful unix shell to help with the task described in the <task> tag. You can run commands and access the filesystem, by using the <stdin> tag to specify input to the shell. Instead of a human user that will interact with you, the user response will be the output from the shell given in the <stdout> command. \n Your current task is this:\n  <task>\n{task_prompt}\n</task>\nIf you think you're done with the entire task., use the <done /> tag as a standalone tag (not inside stdin tag of course). DO NOT JUST RESPOND WITH CODE, THE CODE HAS TO BE INSIDE <stdin> TAGS TO WORK."
-    },
-    {
-        "role": "assistant",
-        "content": "Let's start by looking at the current dir to get a hang of it. \n <stdin>pwd</stdin>"
-    },
-    {
-        "role": "user",
-        "content": f"<stdout>{challenge_workdir}</stdout>"
-    }
-]
-
-def get_response(add_assistant_response=True):
-    global messages
-    print('get response start', messages)
-    response = requests.post(
-    url="https://openrouter.ai/api/v1/chat/completions",
-    headers={
-        "Authorization": f"Bearer {OPENROUTER_API_KEY}",
-    },
-    data=json.dumps({
-        "model": model, #"openai/gpt-3.5-turbo", # Optional'anthropic/claude-3-opus', #
-        "messages": messages
-    })
-    )
-    data = response.json()
-    print('data', data)
-    assert len(data['choices']) == 1, 'Expected one choice'
-    response_text = data['choices'][0]['message']['content']
-    print('res text', response_text)
-    print('get reponse end')
-    if add_assistant_response:
-        messages.append({
-            "role": "assistant",
-            "content": response_text
-        })
-    return response_text
-
-tw = AgentShell(cwd=challenge_workdir)
-max_rounds = 10
-for round_i in range(max_rounds):
-    print('Round', round_i)
-    response_text = get_response()
-    xml_structures = extract_xml_structures(response_text)
-    print('xml structs', xml_structures)
-    #assert len(xml_structures) > 0, 'Expected at least one xml structure'
-    if len(xml_structures) == 0:
-        print('No xml structures found.')
+def start_agent(challenge, model, max_rounds):
+    tw = AgentShell(cwd=challenge.workdir)
+    messages = get_pre_prompt(challenge)
+    for round_i in range(max_rounds):
+        print('Round', round_i)
+        response_text = get_completion(
+            messages, model, OPENROUTER_API_KEY, add_assistant_response=True)
+        print('messages length', len(messages))
+        xml_structures = extract_xml_structures(response_text)
+        print('xml structs', xml_structures)
+        #assert len(xml_structures) > 0, 'Expected at least one xml structure'
+        if len(xml_structures) == 0:
+            print('No xml structures found.')
+            messages.append({
+                "role": "user",
+                "content": "ERROR. No stdin or done tag found. Remember that all code to be executed in shell must be in <stdin> tags. If you think you're done with the task, use the <done /> tag."
+            })
+            continue
+        #assert len(xml_structures) == 1, 'Expected one xml structure'
+        stdout_string = ''
+        for xml_structure in xml_structures:
+            print('xml struct', xml_structure)
+            if xml_structure['tag'] == 'done':
+                print('Agent used <done /> tag to signal completion. Exiting.')
+                break
+            assert xml_structure['tag'] == 'stdin', 'Expected stdin tag'
+            command_input = xml_structure['content']
+            print('command input', command_input)
+            command_id, command_output = tw.round_trip(command_input)
+            print('%', command_id)
+            print('command output', command_output)
+            stdout_string += '<stdout>' + command_output + '</stdout>'
+        print('user will reply:', stdout_string)
         messages.append({
             "role": "user",
-            "content": "ERROR. No stdin or done tag found. Remember that all code to be executed in shell must be in <stdin> tags. If you think you're done with the task, use the <done /> tag."
+            "content": stdout_string
         })
-        continue
-    #assert len(xml_structures) == 1, 'Expected one xml structure'
-    stdout_string = ''
-    for xml_structure in xml_structures:
-        print('xml struct', xml_structure)
-        if xml_structure['tag'] == 'done':
-            print('Agent used <done /> tag to signal completion. Exiting.')
-            break
-        assert xml_structure['tag'] == 'stdin', 'Expected stdin tag'
-        command_input = xml_structure['content']
-        print('command input', command_input)
-        command_id, command_output = tw.round_trip(command_input)
-        print('%', command_id)
-        print('command output', command_output)
-        stdout_string += '<stdout>' + command_output + '</stdout>'
-    print('user will reply:', stdout_string)
-    messages.append({
-        "role": "user",
-        "content": stdout_string
-    })
-    print('Appended stdout')
+        print('Appended stdout')
+    print('end of rounds')
 
-print('END')
+# take command line arguments
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser(description='Run agent for challenge')
+    parser.add_argument('--challenge_id',  type=str, help='Challenge ID')
+    parser.add_argument('--max_rounds', type=int, default=10, help='Max rounds')
+    parser.add_argument('--model', type=str, default='openai/gpt-3.5-turbo', help='Model')
+    args = parser.parse_args()
+    print('challenge id', args.challenge_id)
+    challenge = get_challenge_by_id(args.challenge_id)
+    print('loaded challenge', challenge.challenge_id)
+    print('starting agent')
+    start_agent(challenge, args.model, args.max_rounds)
